Added Login and logout buttons to storybook and small styling changes
